Fixing timestamp in sealed filename

+++ orkbasecxx/AudioTape.cpp
@@ -768,7 +768,7 @@ CStdString AudioTape::get_sealed_file_name()
     m_hour.Format("%.2d", date.tm_hour);
     m_min.Format("%.2d", date.tm_min);
     m_sec.Format("%.2d", date.tm_sec);
-    return "out-"+m_localParty+"-"+m_remoteParty+"-"+m_year+m_month+m_day+"-"+m_hour+m_min+m_sec+"-"+m_beginDate*1000*1000+".wav";
+    return "out-"+m_localParty+"-"+m_remoteParty+"-"+m_year+m_month+m_day+"-"+(unsigned  int)(m_hour+m_min+m_sec)+"-"+m_beginDate*1000*1000+".wav";
 }
 
 void AudioTape::GenerateCaptureFilePathAndIdentifier()
